Flattened linode_api.objects.domain submodule into a file
