--- linode_api/objects/domain/__init__.py
@@ -1,2 +1,0 @@
-from .domain import Domain
-from .record import DomainRecord
